chore: move version-related helper methods to versions file
This change will decouple many packages from utils, specially the common
or defaults one. The utils package is used everywhere, and we want to reduce
the usage
