+++ internal/common/defaults.go
@@ -5,8 +5,8 @@
 package common
 
 import (
+	elasticversion "github.com/elastic/e2e-testing/internal"
 	"github.com/elastic/e2e-testing/internal/shell"
-	"github.com/elastic/e2e-testing/internal/utils"
 	log "github.com/sirupsen/logrus"
 )
 
@@ -80,7 +80,7 @@ func init() {
 // supporting lazy-loading the versions when needed. Basically, the CLI part does not
 // need to load them
 func InitVersions() {
-	v, err := utils.GetElasticArtifactVersion(BeatVersionBase)
+	v, err := elasticversion.GetElasticArtifactVersion(BeatVersionBase)
 	if err != nil {
 		log.WithFields(log.Fields{
 			"error":   err,
@@ -92,7 +92,7 @@ func InitVersions() {
 	BeatVersion = shell.GetEnv("BEAT_VERSION", BeatVersionBase)
 
 	// check if version is an alias
-	v, err = utils.GetElasticArtifactVersion(BeatVersion)
+	v, err = elasticversion.GetElasticArtifactVersion(BeatVersion)
 	if err != nil {
 		log.WithFields(log.Fields{
 			"error":   err,
@@ -110,10 +110,10 @@ func InitVersions() {
 		}).Trace("Beat Version provided: will be used as fallback")
 		fallbackVersion = BeatVersion
 	}
-	BeatVersion = utils.CheckPRVersion(BeatVersion, fallbackVersion)
+	BeatVersion = elasticversion.CheckPRVersion(BeatVersion, fallbackVersion)
 
 	StackVersion = shell.GetEnv("STACK_VERSION", BeatVersionBase)
-	v, err = utils.GetElasticArtifactVersion(StackVersion)
+	v, err = elasticversion.GetElasticArtifactVersion(StackVersion)
 	if err != nil {
 		log.WithFields(log.Fields{
 			"error":   err,
@@ -126,7 +126,7 @@ func InitVersions() {
 	if KibanaVersion == "" {
 		// we want to deploy a released version for Kibana
 		// if not set, let's use StackVersion
-		KibanaVersion, err = utils.GetElasticArtifactVersion(StackVersion)
+		KibanaVersion, err = elasticversion.GetElasticArtifactVersion(StackVersion)
 		if err != nil {
 			log.WithFields(log.Fields{
 				"error":   err,
